MetricBatch: handle /sys/net/quagga/bgp/connage
This string metric is a time duration string in one
of four formats with differing precision and formatting
for equal precisions between them.

Parse and handle it as an IntMetric of seconds. Due to
the reduced precision of long connections, the metric value
will only change once per day by a day for old connections.